// src/enums/filter.enums.ts
export enum FilterDropdownValue {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  ENABLED = 'ENABLED',
  ARCHIVED = 'ARCHIVED',
  PAUSED = 'PAUSED',
  SCHEDULED = 'SCHEDULED',
  RESCHEDULED = 'RESCHEDULED',
  LIVE = 'LIVE',
  COMPLETED = 'COMPLETED',
  ENDED = 'ENDED',
  PROPOSAL = 'PROPOSAL',
  DISABLED = 'DISABLED',
  DELETED = 'DELETED',
  AUTO = 'AUTO',
  MANUAL = 'MANUAL',
  HIGH = '0',
  MEDIUM = '1',
  LOW = '2',
  ALL = 'ALL',
  EXACT = 'EXACT',
  BROAD = 'BROAD',
  PHRASE = 'PHRASE',
  NEGATIVE_EXACT = 'NEGATIVE_EXACT',
  NEGATIVE_PHRASE = 'NEGATIVE_PHRASE',
  NEGATIVE_ASIN_SAME_AS = 'NEGATIVE_ASIN_SAME_AS',
  LEGACY_FOR_SALES = 'LEGACY_FOR_SALES',
  DOWN_ONLY = 'Down Only',
  UP_DOWN = 'Up & Down',
  AUTO_FOR_SALES = 'AUTO_FOR_SALES',
  FIXED_BIDS = 'Fixed Bids',
  FIXED_BIDS_MANUAL = 'MANUAL',
  SP = 'SP',
  SB = 'SB',
  SD = 'SD',
  SV = 'SV',
  ITEM = 'Buy-Box',
  SEARCH = 'Search Ingrid',
  HOMEPAGE = 'Home Page',
  STOCK_UP = 'Stock Up',
  OTHERS = 'Others',
  CATEGORY = 'Category',
  TOPIC = 'Topic',
  BROWSE = 'Browse',
  DESKTOP = 'Desktop',
  MOBILE = 'Mobile',
  APP = 'App',
  BRANDED = 'BRANDED',
  COMPETITOR = 'COMPETITOR',
  GENERIC = 'GENERIC',
  YES = 'Yes',
  NO = 'No',
  PUBLISHED = 'PUBLISHED',
  UNPUBLISHED = 'UNPUBLISHED',
  SYSTEM_PROBLEM = 'SYSTEM_PROBLEM',
  WFS_ELIGIBLE = 'WFS Eligible',
  SELLER_FULFILLED = 'Seller Fulfilled',
  WALMART_FULFILLED = 'Walmart Fulfilled',
  NAME = 'name',
  STATUS = 'state',
  DAILY_BUDGET = 'Budget',
  END_DATE = 'endDate',
  PLACEMENT_BID_MULTIPLIER = 'Placement Bidding',
  AD_GROUP_NAME = 'Adgroup Name',
  AD_GROUP_STATUS = 'Adgroup Status',
  BID = 'bid',
  PRODUCT_AD_STATUS = 'Product Ad status',
  KEYWORD_TARGET_STATUS = `Keyword status`,
  KEYWORD_TARGET_BID = `Keyword bid`,
  BIDDER_SETTING = 'Bidder Setting',
  NEW_KEYWORD = 'New keyword',
  NEW_PRODUCT = 'New Product',
  SELECT_ALL = 'Select All',
  ASIN_SAME_AS = 'ASIN_SAME_AS',
  ASIN_EXPANDED_FROM = 'ASIN_EXPANDED_FROM',
  TARGETING_EXPRESSION = 'TARGETING_EXPRESSION',
  PREDEFINED = 'TARGETING_EXPRESSION_PREDEFINED',
  BUDGET_TYPE_DAILY = WalmartBudgetTypeEnum.DAILY,
  BUDGET_TYPE_TOTAL = WalmartBudgetTypeEnum.TOTAL,
  BUDGET_TYPE_BOTH = WalmartBudgetTypeEnum.BOTH,
  AMAZON = MarketplaceEnum.AMAZON,
  WALMART = MarketplaceEnum.WALMART,
  MARKETPLACE_ALL = MarketplaceEnum.All,
  AMAZON_PROFILE_ID = MetaTypeEnum.AMAZON_PROFILE_ID,
  ANARIX_ACCOUNT_ID = MetaTypeEnum.ANARIX_ACCOUNT_ID,
  AMAZON_ADVERTISER_ID = MetaTypeEnum.AMAZON_ADVERTISER_ID,
  AMAZON_SELLING_PARTNER_ID = MetaTypeEnum.AMAZON_SELLING_PARTNER_ID,
  WALMART_ADVERTISING_ID = MetaTypeEnum.WALMART_ADVERTISING_ID,
  ALL_ACCOUNT = MetaTypeEnum.ALL_ACCOUNT,
  WALMART_PARTNER_ID = MetaTypeEnum.WALMART_PARTNER_ID,
  AD_HOC = ScheduleTypeEnum.AD_HOC,
  DELIVERED = 'Delivered',
  REFUND_COMPLETED = 'REFUND_COMPLETED',
  ORDER_PLACED = 'ORDER_PLACED',
  SHIPPED = 'Shipped',
  REFUND_COMPLETED_MAPPED = 'Refunded',
  ORDER_CANCELLED = 'Cancelled',
  ACKNOWLEDGED = 'Acknowledged',

  CPC = AmazonCostTypeEnum.CPC,
  VCPM = AmazonCostTypeEnum.VCPM,
  REACH = AmazonSDBidOptimizationEnum.REACH,
  CLICKS = AmazonSDBidOptimizationEnum.CLICKS,
  CONVERSIONS = AmazonSDBidOptimizationEnum.CONVERSIONS,
  RULE_BASED = `Amazon's Rule Based`,
  RULE_BASED_MAPPING = `RULE_BASED`,
  METADATA_ERROR = MonitoringMetaDataTypeEnum.ERROR,
  METADATA_SUCCESS = MonitoringMetaDataTypeEnum.SUCCESS,
  METADATA_INFO = MonitoringMetaDataTypeEnum.INFO,
  AMAZON_NA_FULFILLMENT = AmazonFulfillmentTypeEnum.AMAZON_NA,
  AMAZON_DEFAULT_FULFILLMENT = AmazonFulfillmentTypeEnum.DEFAULT,
  AMAZON_NA_FULFILLMENT_MAPPED = AmazonFulfillmentTypeEnum.AMAZON_NA_MAPPED,
  AMAZON_DEFAULT_FULFILLMENT_MAPPED = AmazonFulfillmentTypeEnum.DEFAULT_MAPPED,
  NEW_ITEM = ItemConditionDisplayEnum.NEW,
  REFURBISHED = ItemConditionDisplayEnum.REFURBISHED,
  USED_VERY_GOOD = ItemConditionDisplayEnum.USED_VERY_GOOD,
  USED_ACCEPTABLE = ItemConditionDisplayEnum.USED_ACCEPTABLE,
  USED_POOR = ItemConditionDisplayEnum.USED_POOR,
  USED_LIKE_NEW = ItemConditionDisplayEnum.USED_LIKE_NEW,
  COLLECTIBLE_LIKE_NEW = ItemConditionDisplayEnum.COLLECTIBLE_LIKE_NEW,
  ORDER_PENDING = 'Pending',
  ORDER_RETURNED = 'Returned',
  DRAFT = 'DRAFT',
  DAILY = Frequency.DAILY,
  WEEKLY = Frequency.WEEKLY,
  MONTHLY = Frequency.MONTHLY,
  BIDDER_RULE = RuleTypeEnum.BIDDER_RULE,
  BIDDING_STRATEGY_RULE = RuleTypeEnum.BIDDING_STRATEGY_RULE,
  BUDGET_RULE = RuleTypeEnum.BUDGET_RULE,
  DAYPARTING_RULE = RuleTypeEnum.DAYPARTING_RULE,
  INVENTORY_RULE = RuleTypeEnum.INVENTORY_RULE,
  KEYWORD_HARVESTING_RULE = RuleTypeEnum.KEYWORD_HARVESTING_RULE,
  KEYWORD_RULE = RuleTypeEnum.KEYWORD_RULE,
  PAGE_TYPE_RULE = RuleTypeEnum.PAGE_TYPE_RULE,
  PLACEMENT_RULE = RuleTypeEnum.PLACEMENT_RULE,
  PLATFORM_RULE = RuleTypeEnum.PLATFORM_RULE,
  SOV_RULE = RuleTypeEnum.SOV_RULE,
  STATE_CHANGE_RULE = RuleTypeEnum.STATE_CHANGE_RULE,
  TARGET_RULE = RuleTypeEnum.TARGET_RULE,
  UNKNOWN = RuleTypeEnum.UNKNOWN,
  NONE = 'NONE',
}